Preserve props set after vdom creation
This fixes the issue when the complex prop was not preserved
across connectectedCallback calls.

When complex prop is set as first property then everything
worked as expected. But when complex prop is set after other prop
then it was not stored into this._props and that caused missing it
in the implicit connectedCallback call. Because in connectedCallback
we set props from attributes (simple props) and from this._props.